requirements.txt
pandas==1.4.1
Pillow==8.0.0